Swap Out DB Network Protocol (#198)
* WIP: figuring out grpc analogues for db

* Move ai and db query builders to ahnlich types

* add trait method for create_store

* WIP: implementing dbserver trait

* Mapping predicate conditions to internal types.

* Flesh out pipeline implementation

* add tracing macro to endpoints

* fix clippy lints

---------

+++ ahnlich/grpc_types/src/db/server.rs
@@ -44,13 +44,13 @@ pub struct GetSimN {
 }
 #[derive(Clone, Copy, PartialEq, ::prost::Message)]
 pub struct Del {
-    #[prost(uint32, tag = "1")]
-    pub deleted_count: u32,
+    #[prost(uint64, tag = "1")]
+    pub deleted_count: u64,
 }
 #[derive(Clone, Copy, PartialEq, ::prost::Message)]
 pub struct CreateIndex {
-    #[prost(uint32, tag = "1")]
-    pub created_indexes: u32,
+    #[prost(uint64, tag = "1")]
+    pub created_indexes: u64,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ServerResponse {
@@ -90,8 +90,8 @@ pub mod server_response {
 pub struct StoreInfo {
     #[prost(string, tag = "1")]
     pub name: ::prost::alloc::string::String,
-    #[prost(uint32, tag = "2")]
-    pub len: u32,
+    #[prost(uint64, tag = "2")]
+    pub len: u64,
     #[prost(uint64, tag = "3")]
     pub size_in_bytes: u64,
 }
